Restyle CB a bit, and fix canceling editing
