+++ app/gui/src/project-view/components/GraphEditor.vue
@@ -3,7 +3,6 @@ import {
   codeEditorBindings,
   documentationEditorBindings,
   graphBindings,
-  interactionBindings,
   undoBindings,
 } from '@/bindings'
 import BottomPanel from '@/components/BottomPanel.vue'
@@ -28,7 +27,7 @@ import SceneScroller from '@/components/SceneScroller.vue'
 import TopBar from '@/components/TopBar.vue'
 import { builtinWidgets } from '@/components/widgets'
 import { useDoubleClick } from '@/composables/doubleClick'
-import { keyboardBusy, keyboardBusyExceptIn, unrefElement, useEvent } from '@/composables/events'
+import { keyboardBusy, keyboardBusyExceptIn, unrefElement } from '@/composables/events'
 import { groupColorVar } from '@/composables/nodeColors'
 import type { PlacementStrategy } from '@/composables/nodeCreation'
 import { ActionName, registerHandlers, toggledAction } from '@/providers/action'
@@ -39,7 +38,6 @@ import { provideNodeColors } from '@/providers/graphNodeColors'
 import { provideNodeCreation } from '@/providers/graphNodeCreation'
 import { provideGraphSelection } from '@/providers/graphSelection'
 import { provideStackNavigator } from '@/providers/graphStackNavigator'
-import { provideInteractionHandler } from '@/providers/interactionHandler'
 import { injectKeyboard } from '@/providers/keyboard'
 import { provideWidgetRegistry } from '@/providers/widgetRegistry'
 import type { Node, NodeId } from '@/stores/graph'
@@ -294,32 +292,18 @@ const actionHandlers = registerHandlers({
   ),
 })
 
-// === Interactions ===
-
-const interaction = provideInteractionHandler()
-const interactionBindingsHandler = interactionBindings.handler({
-  cancel: () => interaction.handleCancel(),
-})
-
-useEvent(
-  window,
-  'keydown',
-  (event) =>
-    interactionBindingsHandler(event) ||
-    (!keyboardBusy() && undoBindingsHandler(event)) ||
-    (!keyboardBusy() && graphBindingsHandler(event)) ||
-    (!keyboardBusyExceptIn(codeEditorArea.value) && codeEditorHandler(event)) ||
-    (!keyboardBusyExceptIn(documentationEditorArea.value) && documentationEditorHandler(event)) ||
-    (!keyboardBusy() && graphNavigator.keyboardEvents.keydown(event)),
-)
-
-useEvent(window, 'pointerdown', (e) => interaction.handlePointerEvent(e, 'pointerdown'), {
-  capture: true,
-})
-
-useEvent(window, 'pointerup', (e) => interaction.handlePointerEvent(e, 'pointerup'), {
-  capture: true,
-})
+/**
+ * A keyevent handler which handles shortcuts of this and some more panels.
+ * See also https://github.com/enso-org/enso/issues/10414
+ */
+function broadShortcutHandler(event: KeyboardEvent) {
+  if (!keyboardBusy() && undoBindingsHandler(event)) return
+  if (!keyboardBusy() && graphBindingsHandler(event)) return
+  if (!keyboardBusyExceptIn(codeEditorArea.value) && codeEditorHandler(event)) return
+  if (!keyboardBusyExceptIn(documentationEditorArea.value) && documentationEditorHandler(event))
+    return
+  if (!keyboardBusy() && graphNavigator.keyboardEvents.keydown(event)) return
+}
 
 function tryGetSelectionDocUrl() {
   const selected = nodeSelection.tryGetSoleSelection()
@@ -683,6 +667,7 @@ const contextMenuActions: ActionName[] = [
     :style="groupColors"
     @dragover.prevent
     @drop.prevent="handleFileDrop($event)"
+    @keydown="broadShortcutHandler"
   >
     <div class="vertical">
       <ContextMenuTrigger
